Nothing from the targets subdir used in the upper module
